Last changes

+++ src/Pages/GamePage/index.jsx
@@ -1,5 +1,5 @@
 import "./style.scss";
-import React, { useState, useEffect, useRef, useLayoutEffect } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import { useParams, useHistory } from "react-router-dom";
 import { useSelector } from "react-redux";
 import Cookie from "js-cookie";
@@ -11,18 +11,22 @@ import Countdown from "./Countdown/index";
 import ModalDiv from "./Modal/index";
 import { motion } from "framer-motion";
 import { Prompt } from "react-router-dom";
+import { AudioPlayerProvider } from "react-use-audio-player";
+import PyramidRaceAudio from "./assets/PyramidRaceMusicOGG.ogg";
+import AudioPlayer from "../../Components/AudioPlayer/index.jsx";
 
 const Game = () => {
   let { id } = useParams();
   const userId = useSelector((state) => state.id);
   const tokenCookie = Cookie.get("token");
-  const [game, setGame] = useState({});
+  const [game, setGame] = useState(null);
   const [questions, setQuestions] = useState([]);
   const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0);
   const [currentQuestion, setCurrentQuestion] = useState({});
   const [gameOn, setGameOn] = useState(false);
   const [newQuestionTime, setNewQuestionTime] = useState(new Date(Date.now()));
   const [currentStep, setCurrentStep] = useState(0);
+  const [opponentStep, setOpponentStep] = useState(0);
   const [modalIsOpen, setIsOpen] = useState(false);
   const [gameHistories, setGameHistories] = useState([]);
   const [perso1animation, setPerso1Animation] = useState({ x: 0, y: 0 });
@@ -30,11 +34,14 @@ const Game = () => {
   const [firstGameHistory, setFirstGameHistory] = useState({});
   const history = useHistory();
   const pyramidRef = useRef();
-  const [count, setCount] = useState(0);
   const [count2, setCount2] = useState(0);
   const [count3, setCount3] = useState(0);
   const [timePlayer1, setTimePlayer1] = useState(0);
   const [timePlayer2, setTimePlayer2] = useState(0);
+
+  // Status possible: pending, gameFetched, questionsFetched, gameReady, gameOver
+
+  const status = useRef("pending");
 
   const openModal = () => {
     setIsOpen(true);
@@ -46,12 +53,12 @@ const Game = () => {
   };
 
   useEffect(() => {
-    if (questions.length === 12) {
+    if (status.current === "questionsFetched") {
       setCurrentQuestion(questions[currentQuestionIndex]);
-      console.log("game oooooooooooooooooooooooooon");
       setTimePlayer1(Date.now());
       setCurrentQuestionIndex(currentQuestionIndex + 1);
       setGameOn(true);
+      status.current = "gameReady";
     }
   }, [questions]);
 
@@ -60,28 +67,26 @@ const Game = () => {
   }, []);
 
   useEffect(() => {
-    if (game && count === 2) {
+    if (game && status.current === "gameFetched") {
       if (userId === game.player2_id) {
         fetchHistoryPlayer1();
       }
       fetchQuestions();
     }
-    // return () => {
-    //   if (game && count === 2) {
-    //     console.log("test usablecallback");
-    //     userId === game.player1_id ? destroyGame() : forfeitGame();
-    //   }
-    // };
-  }, [count, game]);
-
-  useEffect(() => {
-    setCount(count + 1);
+    return () => {
+      if (game && status.current === "gameReady") {
+        userId === game.player1_id ? destroyGame() : forfeitGame();
+      }
+    };
   }, [game]);
 
   const fetchGame = () => {
     fetch(`https://pyramid-race-api.herokuapp.com/games/${id}`)
       .then((response) => response.json())
-      .then((data) => setGame(data))
+      .then((data) => {
+        setGame(data);
+        status.current = "gameFetched";
+      })
       .catch((error) => console.log(error));
   };
 
@@ -90,7 +95,11 @@ const Game = () => {
       `https://opentdb.com/api.php?amount=12&category=${game.category}&difficulty=${game.difficulty}&type=multiple`
     )
       .then((response) => response.json())
-      .then((data) => setQuestions(data.results));
+      .then((data) => {
+        setQuestions(data.results);
+        status.current = "questionsFetched";
+      })
+      .catch((error) => console.log(error));
   };
 
   const fetchHistoryPlayer1 = () => {
@@ -105,29 +114,22 @@ const Game = () => {
 
   const gameEnd = (firstWinnerId) => {
     let winner_id;
-    if (!firstWinnerId) {
-      let player1_correct_answers = gameHistories.filter(
-        (game_history) => game_history.response_correct === true
-      ).length;
-      let player1_wrong_answers = gameHistories.filter(
-        (game_history) => game_history.response_correct !== true
-      ).length;
-      let player1_step = player1_correct_answers - player1_wrong_answers;
-      if (player1_step > currentStep) {
-        winner_id = game.player1_id;
-      } else if (player1_step < currentStep) {
-        winner_id = game.player2_id;
-      } else if (player1_step === currentStep) {
-        if (timePlayer1 >= timePlayer2) {
-          winner_id = game.player2_id;
-        } else if (timePlayer1 < timePlayer2) {
-          winner_id = game.player1_id;
-        }
-      }
-    } else {
+    if (firstWinnerId) {
       winner_id = firstWinnerId;
+    } else if (currentStep > opponentStep) {
+      winner_id = game.player2_id;
+    } else if (currentStep < opponentStep) {
+      winner_id = game.player1_id;
+    } else if (opponentStep === currentStep) {
+      if (timePlayer1 >= timePlayer2) {
+        winner_id = game.player2_id;
+      } else if (timePlayer1 < timePlayer2) {
+        winner_id = game.player1_id;
+      }
     }
-
+    console.log("currentStep", currentStep);
+    console.log("opponentStep ", opponentStep);
+    console.log("winner_id", winner_id);
     const data = {
       game: {
         winner_id: winner_id,
@@ -142,12 +144,13 @@ const Game = () => {
       },
       body: JSON.stringify(data),
     })
-      .then((response) => {
+      .then(() => {
         if (userId === winner_id) {
           history.push(`/games/${id}/victory`);
-        } else if (userId != winner_id) {
+        } else if (userId !== winner_id) {
           history.push(`/games/${id}/defeat`);
         }
+        status.current = "gameOver";
       })
       .catch((error) => console.log(error));
   };
@@ -165,7 +168,11 @@ const Game = () => {
         "Content-Type": "application/json",
       },
       body: JSON.stringify(data),
-    }).catch((error) => console.log(error));
+    })
+      .then(() => {
+        status.current = "gameOver";
+      })
+      .catch((error) => console.log(error));
   };
 
   const nextQuestion = (answer_choice, correct_answer) => {
@@ -199,7 +206,6 @@ const Game = () => {
     ) {
       setCurrentStep(currentStep + 1);
       setGameOn(false);
-      console.log("game ooooooooooooooof");
       setCurrentQuestion({});
       if (userId === game.player1_id) {
         nextTurn();
@@ -220,7 +226,6 @@ const Game = () => {
       setNewQuestionTime(new Date(Date.now()));
     } else {
       setGameOn(false);
-      console.log("game ooooof 22222");
       setCurrentQuestion({});
       setCurrentQuestionIndex("");
       if (userId === game.player1_id) {
@@ -229,27 +234,23 @@ const Game = () => {
       } else if (userId === game.player2_id) {
         let totalTime = Date.now() - timePlayer1;
         setTimePlayer1(totalTime);
+        gameEnd();
       }
     }
   };
-  useEffect(() => {
-    console.log(gameOn);
-  }, [gameOn]);
 
   useEffect(() => {
-    if (count3 === 2) {
+    if (count3 === 3) {
       gameEnd();
     }
     setCount3(count3 + 1);
   }, [timePlayer1]);
 
-  const alertUser = (e) => {
-    e.preventDefault();
-    e.returnValue = "";
-  };
-
   const destroyGame = () => {
-    console.log("test destroy");
+    if (status.current === "gameOver") {
+      return;
+    }
+
     fetch(`https://pyramid-race-api.herokuapp.com/games/${id}`, {
       method: "delete",
       headers: {
@@ -260,10 +261,9 @@ const Game = () => {
   };
 
   const forfeitGame = () => {
-    // if (!gameOn) {
-    //   return;
-    // }
-    console.log("test forfeit");
+    if (status.current === "gameOver") {
+      return;
+    }
 
     const data = {
       game: {
@@ -315,7 +315,6 @@ const Game = () => {
           y: -5 * marchHeight - groundHeight,
         });
         setGameOn(false);
-        console.log("game ooooooooooooooooof 3");
         gameEnd(game.player1_id);
       }
     }
@@ -339,16 +338,19 @@ const Game = () => {
       setTimeout(function () {
         if (game_history.response_correct) {
           step += 1;
+          setOpponentStep(opponentStep + 1);
         } else if (step > 0 && !game_history.response_correct) {
           step -= 1;
         }
         if (step > 0) {
           movePlayer1(-step);
+          setOpponentStep(step);
         } else if (step === 0) {
           setPerso2Animation({
             x: 0,
             y: -groundHeight,
           });
+          setOpponentStep(step);
         }
       }, Date.parse(game_history.response_time) - startOpponentGame);
     });
@@ -359,13 +361,8 @@ const Game = () => {
   }, [currentStep]);
 
   useEffect(() => {
-    console.log(perso1animation);
-  }, [perso1animation]);
-
-  useEffect(() => {
-    console.log(gameHistories);
     setCount2(count2 + 1);
-    if (count2 === 1) {
+    if (count2 === 1 && game.turn === "player2") {
       movePlayer2();
     }
   }, [firstGameHistory]);
@@ -379,26 +376,37 @@ const Game = () => {
 
   return (
     <div className="game_page">
-      {userId === game.player1_id && gameOn && (
+      {game && userId === game.player1_id && gameOn && (
         <Prompt
           message={() => "Si vous quittez cette page la partie sera perdue !"}
         />
       )}
-      {userId === game.player2_id && gameOn && (
+      {game && userId === game.player2_id && gameOn && (
         <Prompt
           message={() =>
             "Si vous quittez cette page, vous serez automatiquement déclaré forfait !"
           }
         />
       )}
-
+      <AudioPlayerProvider>
+        <AudioPlayer file={PyramidRaceAudio} />
+      </AudioPlayerProvider>
       <ModalDiv
         modalIsOpen={modalIsOpen}
         closeModal={closeModal}
         step={currentStep}
+        controlsList="play"
       />
-      {((gameOn && userId === game.player2_id && game.turn === "player2") ||
-        (gameOn && userId === game.player1_id && game.turn === "player1")) && (
+      {((status.current === "gameReady" &&
+        game &&
+        gameOn &&
+        userId === game.player2_id &&
+        game.turn === "player2") ||
+        (status.current === "gameReady" &&
+          game &&
+          gameOn &&
+          userId === game.player1_id &&
+          game.turn === "player1")) && (
         <>
           <Countdown onExpire={nextQuestion} resetTick={currentQuestionIndex} />
           <QuestionCard
